+++ packages/tinydoc-plugin-js/lib/__tests__/index--integration.test.js
@@ -15,7 +15,7 @@ describe("[Integration] tinydoc-plugin-js", function() {
       plugins: [
         Subject({
           verbose: false,
-          routeName: 'js-test',
+          routeName: 'test',
           source: TinyTestUtils.tempPath('lib/**/*.js')
         })
       ]
@@ -59,47 +59,11 @@ describe("[Integration] tinydoc-plugin-js", function() {
     tiny.run(function(err, stats) {
       if (err) { return done(err); }
 
-      assert.equal(stats['js:js-test'].count, 3);
-      assert.equal(stats['js:js-test'].modules.count, 2);
-      assert.equal(stats['js:js-test'].entities.count, 1);
+      assert.equal(stats['js:test'].count, 3);
+      assert.equal(stats['js:test'].modules.count, 2);
+      assert.equal(stats['js:test'].entities.count, 1);
 
       done();
-    });
-  });
-
-  describe('@config.readme', function() {
-    it('renders a README file', function(done) {
-      TinyTestUtils.createFile(multiline(function() {;
-        // # Hello JS plugin!
-        //
-        // Welcome.
-      }, true), 'doc/README.md');
-
-      config.plugins = [
-        Subject({
-          verbose: false,
-          routeName: 'js-test',
-          source: TinyTestUtils.tempPath('lib/**/*.js'),
-          readme: 'doc/README.md'
-        })
-      ];
-
-      var tiny = tinydoc(config, {
-        scan: true,
-        write: true,
-        index: true,
-        render: true,
-        stats: true,
-        purge: true
-      });
-
-      tiny.run(function(err, stats) {
-        if (err) { return done(err); }
-
-        assert.ok(stats['js:test'].hasReadme);
-
-        done();
-      });
     });
   });
 });