+++ packages/tinydoc-plugin-js/ui/index.js
@@ -7,16 +7,12 @@ const PreviewHandler = require('./PreviewHandler');
 
 Storage.register(K.CFG_CLASS_BROWSER_SHOW_PRIVATE, false);
 
+OutletManager.define('CJS::Landing', { firstMatching: true });
 OutletManager.define('CJS::ModuleBody');
-OutletManager.define('CJS::ModuleHeader::Type', {
-  firstMatching: true
-});
-
+OutletManager.define('CJS::ModuleHeader::Type', { firstMatching: true });
 OutletManager.define('CJS::Tag');
 OutletManager.define('CJS::ExampleTags');
-OutletManager.define('CJS::ExampleTag', {
-  firstMatching: true
-});
+OutletManager.define('CJS::ExampleTag', { firstMatching: true });
 
 tinydoc.use(function CJSPlugin(api) {
   const configs = tinydoc.getRuntimeConfigs('cjs');