+++ packages/tinydoc-plugin-static/ui/components/StandaloneStaticFile.js
@@ -1,0 +1,15 @@
+const React = require('react');
+const Document = require('components/Document');
+const StaticFile = require('./StaticFile');
+
+const StandaloneStaticFile = React.createClass({
+  render() {
+    return (
+      <Document>
+        <StaticFile {...this.props} />
+      </Document>
+    );
+  }
+});
+
+module.exports = StandaloneStaticFile;